requirements_pip.txt
django
psycopg2
cx_Oracle<7
django-leaflet